refactor Socks to utilities
